previous activity restructure
