+++ web/src/pages/apd/previous-activities/ApdPreviousActivityTableMMIS.test.js
@@ -14,7 +14,7 @@ import {
   setPreviousActivityFederalActualExpenseforMMIS50FFP,
   setPreviousActivityFederalActualExpenseforMMIS75FFP,
   setPreviousActivityFederalActualExpenseforMMIS90FFP
-} from '../actions/editApd';
+} from '../../../actions/editApd';
 
 describe('apd previous activity table, mmis component', () => {
   const props = {
