+++ web/src/pages/apd/previous-activities/PreviousActivities.js
@@ -6,13 +6,13 @@ import { titleCase } from 'title-case';
 import ApdPreviousActivityTableHI from './ApdPreviousActivityTable';
 import ApdPreviousActivityTableMMIS from './ApdPreviousActivityTableMMIS';
 import ApdPreviousActivityTableTotal from './ApdPreviousActivityTableTotal';
-import Waypoint from './ConnectedWaypoint';
-import { setPreviousActivitySummary } from '../actions/editApd';
-import RichText from '../components/RichText';
-import { Section, Subsection } from '../components/Section';
-import { t } from '../i18n';
-import { selectPreviousActivitySummary } from '../reducers/apd.selectors';
-import AlertMissingFFY from '../components/AlertMissingFFY';
+import Waypoint from '../../../containers/ConnectedWaypoint';
+import { setPreviousActivitySummary } from '../../../actions/editApd';
+import RichText from '../../../components/RichText';
+import { Section, Subsection } from '../../../components/Section';
+import { t } from '../../../i18n';
+import { selectPreviousActivitySummary } from '../../../reducers/apd.selectors';
+import AlertMissingFFY from '../../../components/AlertMissingFFY';
 
 const PreviousActivities = ({ previousActivitySummary, setSummary }) => {
   const onChange = value => setSummary(value);
